\name{pspearman-package}
\alias{pspearman}
\docType{package}
\title{Spearman's rank correlation test (package)}
\description{
  Currently, the only function available in this package is
  the function \code{spearman.test}, which is a modification
  of the function \code{cor.test(method="spearman")} improving
  the accuracy of the p-value for small values of \code{n}.
}
\seealso{
  \code{\link{spearman.test}}
}
\keyword{ package }

